feat: WebAssembly package for versatile integration
