+++ packages/modular-indexer-light/index.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Testing Page | Modular Indexer (Light)</title>
+    <script defer type="module" src="./wasm_exec.js"></script>
+    <script defer type="module" src="./index.js"></script>
+    <script defer type="module">
+        import {lightLoadWasm} from "./index.js";
+
+        // Load the Wasm module from the web.
+        await lightLoadWasm();
+
+        // Then the following APIs are globals.
+        window.lightSetConfig({
+            "committeeIndexers": {
+                "s3": [
+                    {
+                        "region": "us-west-2",
+                        "bucket": "nubit-modular-indexer-brc-20",
+                        "name": "nubit-official-00"
+                    }
+                ],
+                "da": []
+            },
+            "verification": {
+                "bitcoinRPC": "https://bitcoin-mainnet-archive.allthatnode.com",
+                "metaProtocol": "brc-20",
+                "minimalCheckpoint": 1
+            },
+        });
+        window.lightInitialize();
+
+        async function getBlockHeight() {
+            document.getElementById("block-height").textContent = `${await window.lightGetBlockHeight()}`;
+        }
+    </script>
+</head>
+<body>
+<div>
+    <input id="btn-block-height" type="button" onclick="getBlockHeight()" value="Block height">
+    <label for="btn-block-height" id="block-height">(unknown)</label>
+</div>
+</body>
+</html>
